+++ assets/javascripts/kitten/components/headers/header-titles.js
@@ -13,11 +13,10 @@ export class HeaderTitles extends React.Component {
   }
 
   render() {
-    const { title, subtitle, className, ...other } = this.props
+    const { title, subtitle, ...other } = this.props
 
     return (
-      <HeaderItem { ...other }
-                  className={ classNames('k-Header__titles', className) }>
+      <HeaderItem className='k-Header__titles' { ...other }>
         <p className="k-Header__title">{ title }</p>
         { this.renderSubtitle() }
       </HeaderItem>